Upscale Script/requirements.txt
numpy==1.24.3
Pillow==10.4.0
spandrel==0.4.0
spandrel_extra_arches==0.2.0
torch==2.4.1+cu124
tqdm==4.66.5
